Fix Date examples in README.md
It looks like the date examples in the README are incorrect or at the very least not following the convention of not mutating the original value in transforms. I believe the issues is that instead of re-assigning value to the moment object, the intention was to allocate a new variable by the name of date with the moment object.

Let me know if this is correct. Or if I'm suppose to be following the pattern described in the read me and I'm missing something.