Changes array.ensure to work on the original value
This fixes to match the documentation, which says that a single element
should get cast to an array containing that element. Previously the
transform added in the constructor would convert non-arrays to null
before the ensure transform function would see them.

Fixes #343

WARNING: This is a breaking change. `ensure` now ignores any
transformations added before it, including the transform in the
constructor that parses JSON.