Update i18n doc
Currently doc is confusing. The pattern of calling  `setLocale` with hard written text or even `t('foobar')` is NOT meant for i18n, as it doesn't support language switching for instance.

The second pattern with using function that returns i18n objects (`{key:'foobar', values: {min}`) is the only right pattern for i18n.